refactored imaging module plus tests into package
