+++ autoarray/inversion/inversion/factory.py
@@ -1,7 +1,7 @@
 from typing import Dict, List, Optional, Union
 
-from autoarray.dataset.imaging import Imaging
-from autoarray.dataset.imaging import WTildeImaging
+from autoarray.dataset.imaging.imaging import Imaging
+from autoarray.dataset.imaging.w_tilde import WTildeImaging
 from autoarray.dataset.interferometer import WTildeInterferometer
 from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.visibilities import Visibilities
