+++ test_autoarray/dataset/imaging/files/3x3_ones_central_two.fits
