+++ test_autoarray/dataset/imaging/files/3x3_threes.fits
