+++ test_autoarray/dataset/imaging/files/3x3_twos.fits
