+++ test_autoarray/dataset/imaging/files/4x3_ones.fits
