+++ test_autoarray/dataset/imaging/files/positions.json
